Use extensions.py file
Now we have a better organization of the files and the app is more
modular
