+++ APITaxi/__init__.py
@@ -11,28 +11,12 @@ from flask_bootstrap import Bootstrap
 import os
 from flask.ext.redis import FlaskRedis
 from flask.ext.login import user_logged_out
-from .utils.redis_geo import GeoRedis
-redis_store = FlaskRedis.from_custom_provider(GeoRedis)
-from dogpile.cache import make_region
-def user_key_generator(namespace, fn, **kw):
-    def generate_key(*args, **kwargs):
-        return fn.__name__ +\
-             "_".join(str(s) for s in args) +\
-             "_".join(k+"_"+str(v) for k,v in kwargs.iteritems())
-    return generate_key
-region_users = make_region('users', function_key_generator=user_key_generator)
-from .models import db
 from flask.ext.restplus import abort
-from flask.ext.uploads import (UploadSet, configure_uploads,
-            DOCUMENTS, DATA, ARCHIVES, IMAGES)
 from slacker import Slacker
 from .utils.request_wants_json import request_wants_json
 from sqlalchemy import distinct
 from rtree import index
-from .index_zupc import IndexZUPC
-from .utils.request_wants_json import request_wants_json
-region_taxi = make_region('taxis')
-region_hails = make_region('hails')
+from .extensions import *
 
 valid_versions = ['1', '2']
 def check_version(sender, **extra):
@@ -53,11 +37,6 @@ def add_version_header(sender, response, **extra):
 def invalidate_user(sender, user, **extra):
     region_users.invalidate(user)
 
-
-documents = UploadSet('documents', DOCUMENTS + DATA + ARCHIVES)
-images = UploadSet('images', IMAGES)
-
-index_zupc = IndexZUPC()
 
 def create_app(sqlalchemy_uri=None):
     app = Flask(__name__)
