+++ APITaxi/index_zupc.py
@@ -2,7 +2,7 @@
 from operator import itemgetter
 from rtree import index
 from sqlalchemy import distinct
-from . import db
+from .extensions import db
 from functools import wraps
 from werkzeug.wrappers import Response
 from shapely.geometry import Point
