+++ APITaxi/models/__init__.py
@@ -1,3 +1,1 @@
 # -*- coding: utf-8 -*-
-from flask_sqlalchemy import SQLAlchemy
-db = SQLAlchemy(session_options={"autoflush":False})
