+++ APITaxi/utils/model_form.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from flask.ext.wtf import Form
 from wtforms_alchemy import model_form_factory
-from ..models import db
+from .. import db
 
 BaseModelForm = model_form_factory(Form)
